move enterprise/internal/cody to internal/cody
