+++ enterprise/cmd/frontend/internal/embeddings/resolvers/BUILD.bazel
@@ -13,10 +13,10 @@ go_library(
         "//cmd/frontend/backend",
         "//cmd/frontend/graphqlbackend",
         "//cmd/frontend/graphqlbackend/graphqlutil",
-        "//enterprise/internal/cody",
         "//internal/api",
         "//internal/auth",
         "//internal/authz",
+        "//internal/cody",
         "//internal/conf",
         "//internal/database",
         "//internal/embeddings",
