+++ enterprise/cmd/frontend/shared/BUILD.bazel
@@ -35,7 +35,6 @@ go_library(
         "//enterprise/cmd/frontend/internal/repos/webhooks",
         "//enterprise/cmd/frontend/internal/search",
         "//enterprise/cmd/frontend/internal/searchcontexts",
-        "//enterprise/internal/cody",
         "//enterprise/internal/oobmigration/migrations",
         "//enterprise/internal/scim",
         "//internal/codeintel",
