--- enterprise/internal/cody/BUILD.bazel
@@ -1,37 +1,0 @@
-load("//dev:go_defs.bzl", "go_test")
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "cody",
-    srcs = ["feature_flag.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/internal/cody",
-    visibility = ["//:__subpackages__"],
-    deps = [
-        "//cmd/frontend/backend",
-        "//cmd/frontend/envvar",
-        "//enterprise/internal/licensing",
-        "//internal/actor",
-        "//internal/auth",
-        "//internal/conf",
-        "//internal/conf/deploy",
-        "//internal/database",
-        "//internal/featureflag",
-        "//lib/errors",
-        "@com_github_sourcegraph_log//:log",
-    ],
-)
-
-go_test(
-    name = "cody_test",
-    srcs = ["feature_flag_test.go"],
-    embed = [":cody"],
-    deps = [
-        "//enterprise/internal/licensing",
-        "//internal/actor",
-        "//internal/conf",
-        "//internal/conf/deploy",
-        "//internal/featureflag",
-        "//lib/errors",
-        "//schema",
-    ],
-)
