+++ internal/codycontext/BUILD.bazel
@@ -6,8 +6,8 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/internal/codycontext",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//enterprise/internal/database",
         "//internal/api",
+        "//internal/database",
         "//internal/embeddings",
         "//internal/embeddings/embed",
         "//internal/metrics",
